Calculate the influence curve based SE with given fitted HAL object and evaluation points

Calculate the influence curve based SE with given fitted HAL object and evaluation points

Usage

IC_based_se(
  X,
  Y,
  hal_fit,
  eval_points,
  family = "binomial",
  X_unpenalized = NULL
)

Arguments

X

An input matrix with dimensions number of observations -by- number of covariates that will be used to derive the design matrix of basis functions.

Y

A numeric vector of observations of the outcome variable.

hal_fit

The HAL fit object.

family

A character or a family object (supported by glmnet) specifying the error/link family for a generalized linear model. See more details in fit_hal.

X_unpenalized

TAn input matrix with dimensions number of observations -by- number of covariates that are not penalized when fit HAL.

Details

The procedure calculates the SE with the following steps: 1). get all the selected basis functions and corresponding from hal_fit 2). calculate influence curves for coefficients 3). use delta method to calculate influence curves for points on the curve