fit undersoomthed HAL function with global criterion

fit undersoomthed HAL function with global criterion

Usage

undersmooth_hal(
  X,
  Y,
  fit_init,
  Nlam = 20,
  family = c("gaussian", "binomial", "poisson", "cox")
)

Arguments

X

An input matrix with dimensions number of observations -by- number of covariates that will be used to derive the design matrix of basis functions.

Y

A numeric vector of observations of the outcome variable.

fit_init

The initial HAL fit object from the output list of undersmooth_init.

Nlam

Number of lambda candidates. The sequence ranges from fit_init$lambda_star to fit_init$lambda_star*10^(-3).

family

A character or a family object (supported by glmnet) specifying the error/link family for a generalized linear model. See more details in fit_hal.

Details

The procedure select the lambda that satisfies the global undersmoothing criterion (\( P_n(\phi_{s,i}(Y-\bar{Q}_{n,\lambda}))\leq \freq{\sigma_n}{\sqrt{n}log(n)} \)). It performs the following steps: 1). get all the basis functions from cv-hal 2). calculate a grid of new lambdas by scaling the cv-lambda from cv-hal with seq(from=0, to=-3, length=Nlam) 3). refit lasso using glmnet) with all the basis functions and with the grid of new lambdas 4).calculate the scores (\( P_n(\phi_{s,i}(Y-\bar{Q}_{n,\lambda})) \))of each lasso fit 5).find the biggest lambda that max score (\( \leq \freq{\sigma_n}{\sqrt{n}log(n)} \))